{"schema":"decodex/commit/1","summary":"Split acceptance and benchmark helpers into modules","authority":"manual"}
