+++ packages/elf-service/tests/acceptance/chunk_search/tests_helpers/context.rs
@@ -1,0 +1,59 @@
+use crate::acceptance;
+use elf_service::{ElfService, Providers};
+use elf_testkit::TestDatabase;
+
+pub(in super::super) struct TestContext {
+	pub(in super::super) service: ElfService,
+	pub(in super::super) test_db: TestDatabase,
+	pub(in super::super) embedding_version: String,
+}
+
+pub(in super::super) async fn setup_context(
+	test_name: &str,
+	providers: Providers,
+) -> Option<TestContext> {
+	let Some(test_db) = acceptance::test_db().await else {
+		eprintln!("Skipping {test_name}; set ELF_PG_DSN to run this test.");
+
+		return None;
+	};
+	let Some(qdrant_url) = acceptance::test_qdrant_url() else {
+		eprintln!("Skipping {test_name}; set ELF_QDRANT_URL to run this test.");
+
+		return None;
+	};
+	let collection = test_db.collection_name("elf_acceptance");
+	let docs_collection = test_db.collection_name("elf_acceptance_docs");
+	let cfg = acceptance::test_config(
+		test_db.dsn().to_string(),
+		qdrant_url,
+		4_096,
+		collection,
+		docs_collection,
+	);
+	let service =
+		acceptance::build_service(cfg, providers).await.expect("Failed to build service.");
+
+	acceptance::reset_db(&service.db.pool).await.expect("Failed to reset test database.");
+
+	reset_collection(&service).await;
+
+	let embedding_version = format!(
+		"{}:{}:{}",
+		service.cfg.providers.embedding.provider_id,
+		service.cfg.providers.embedding.model,
+		service.cfg.storage.qdrant.vector_dim
+	);
+
+	Some(TestContext { service, test_db, embedding_version })
+}
+
+pub(in super::super) async fn reset_collection(service: &ElfService) {
+	acceptance::reset_qdrant_collection(
+		&service.qdrant.client,
+		&service.qdrant.collection,
+		service.qdrant.vector_dim,
+	)
+	.await
+	.expect("Failed to reset Qdrant collection.");
+}
